Renaming xmlforms for analytics and grouping replication functions under medic
